modify : 사운드 적용 위치 변경 (main -> app)
