+++ 2-Project/GameApp/src/resource/js/app.js
@@ -1,4 +1,5 @@
 import * as Utils from './utils/constant.js';
+import * as Sound from './sound.js';
 import Popup from './popup.js';
 import Field from './field.js';
 
@@ -10,12 +11,6 @@ const gameMsgElement = popupElement.querySelector('.game-massage');
 const playBtnElement = popupElement.querySelector('.play');
 const replayBtnElement = popupElement.querySelector('.replay');
 const countElement = root.querySelector('.game-count .count');
-
-const carrotSound = new Audio('./resource/sound/carrot_pull.mp3');
-const alertSound = new Audio('./resource/sound/alert.wav');
-const bugSound = new Audio('./resource/sound/bug_pull.mp3');
-const bgSound = new Audio('./resource/sound/bg.mp3');
-const winSound = new Audio('./resource/sound/game_win.mp3');
 
 let gameStatus = false;
 let countTimer = null;
@@ -48,8 +43,8 @@ const setReverseCountTIme = time => {
       timerElement.innerHTML = `0.00`;
       gamePopup.show();
       gamePopup.showPopupText('replay game?');
-      stopSound(bgSound);
-      playSound(alertSound);
+      Sound.stopBg();
+      Sound.playAlert();
       return clearInterval(countTimer);
     }
     timerElement.innerHTML = millisec;
@@ -66,16 +61,16 @@ const initialGame = () => {
 
 const startGame = () => {
   initialGame();
-  playSound(bgSound);
+  Sound.playBg();
   setReverseCountTIme(Utils.SET_TIMER);
 };
 
 const finishGame = msg => {
   initialGame();
-  stopSound(bgSound);
-  playSound(alertSound);
+  Sound.playBg();
+  Sound.playAlert();
   if (gameStatus) {
-    playSound(winSound);
+    Sound.playWin();
     gameStatus = !gameStatus;
   }
   gamePopup.showPopupText(msg);
@@ -84,7 +79,7 @@ const finishGame = msg => {
 
 const removeCount = () => {
   itemCount--;
-  playSound(carrotSound);
+  Sound.playCarrot();
   if (itemCount === 0) {
     gameStatus = !gameStatus;
     return finishGame('Congratulations!! 👏');
@@ -92,15 +87,5 @@ const removeCount = () => {
   return (countElement.innerHTML = itemCount);
 };
 
-const playSound = sound => {
-  sound.currentTime = 0;
-  return sound.play();
-};
-
-const stopSound = sound => {
-  return sound.pause();
-};
-
 gamePopup.setClickListener(startGame);
 gameField.setClickListener(onClickField);
-// fieldElement.addEventListener('click', onClickFieldTarget);
